refactor(core): split build-candidates and bar/rule/text geometry
Extract candidate identity index and stat/annotation datum resolver from
buildPipelineCandidates; split rects, segments, and glyphs builders out of
the bar-rule-text facade. Characterization tests pin source-backed vs
aggregate lineages and text/rule batch kinds.
