+++ packages/core/src/pipeline/geometry-glyphs.ts
@@ -1,0 +1,70 @@
+/**
+ * Text glyph geometry batch builder.
+ */
+import type { GlyphsBatch } from "../scene.js";
+import { bandKey } from "../scales/train.js";
+
+import type { LayerFrame, PipelineWarning, ResolvedColorScale } from "./types.js";
+import { colorOf } from "./types.js";
+import type { Frame } from "./geometry-shared.js";
+import { DEFAULT_TEXT_SIZE, positionOf, removedWarning } from "./geometry-shared.js";
+
+export function glyphsBatch(
+  frame: LayerFrame,
+  fx: Frame,
+  color: ResolvedColorScale | null,
+  warnings: PipelineWarning[],
+): GlyphsBatch | null {
+  const { binding, n } = frame;
+  const params = (binding.layer.params ?? {}) as {
+    anchor?: "start" | "middle" | "end";
+    size?: number;
+    dx?: number;
+    dy?: number;
+    alpha?: number;
+  };
+  const dx = params.dx ?? 0;
+  const dy = params.dy ?? 0;
+  const positions: number[] = [];
+  const rowIndex: number[] = [];
+  const texts: string[] = [];
+  const colors: string[] = [];
+  const wantsColors =
+    color !== null && (frame.colorValues !== null || binding.color.scaledConstant !== null);
+  let removed = 0;
+  for (let row = 0; row < n; row++) {
+    const tx = positionOf(fx.xScale, frame.xNumeric, frame.xValues, row, frame.offsetX);
+    const ty = positionOf(fx.yScale, frame.yNumeric, null, row, frame.offsetY);
+    const label =
+      binding.labelConstant ?? (frame.labelValues === null ? null : frame.labelValues[row]);
+    if (Number.isNaN(tx) || Number.isNaN(ty) || label === null) {
+      removed++;
+      continue;
+    }
+    positions.push(tx * fx.innerWidth + dx, fx.innerHeight - ty * fx.innerHeight + dy);
+    rowIndex.push(frame.rowIndex[row]!);
+    texts.push(bandKey(label));
+    if (wantsColors) {
+      const value =
+        frame.colorValues === null ? binding.color.scaledConstant! : frame.colorValues[row]!;
+      colors.push(colorOf(color, value));
+    }
+  }
+  removedWarning(removed, binding.index, warnings);
+  if (texts.length === 0) return null;
+
+  const batch: GlyphsBatch = {
+    kind: "glyphs",
+    layerIndex: binding.index,
+    panelIndex: 0,
+    positions: Float32Array.from(positions),
+    rowIndex: Uint32Array.from(rowIndex),
+    texts,
+    color: binding.color.constant,
+    size: params.size ?? DEFAULT_TEXT_SIZE,
+    anchor: params.anchor ?? "middle",
+    alpha: params.alpha ?? 1,
+  };
+  if (wantsColors) batch.colors = colors;
+  return batch;
+}
